fix: amount to int

+++ migrations/000003_create_expenses_table.up.sql
@@ -3,7 +3,7 @@ CREATE TABLE expenses
     id          SERIAL PRIMARY KEY,
     user_id     INT            NOT NULL REFERENCES users (id) ON DELETE CASCADE,
     category_id INT,
-    amount      NUMERIC(12, 2) NOT NULL,
+    amount      INT NOT NULL,
     description TEXT,
     created_at  TIMESTAMP DEFAULT NOW(),
     updated_at  TIMESTAMP DEFAULT NOW()
