+++ migrations/000004_create_add_income_function.down.sql
@@ -1,1 +1,1 @@
-DROP FUNCTION IF EXISTS add_income(user_id INT, category_id INT, amount NUMERIC, description TEXT);
+DROP FUNCTION IF EXISTS add_income(user_id INT, category_id INT, amount INT, description TEXT);
